# Benchmark script that also calculates CPU->GPU speedup.