--- a/benchmark/benchmarks.jl
+++ b/benchmark/benchmarks.jl
@@ -40,3 +40,23 @@ for N in Ns
     t64  = TimerOutputs.time(timer[bn64])
     @printf("%s: %.3f\n", benchmark_name(N), t64/t32)
 end
+
+Oceananigans.@hascuda begin
+    println("\nGPU Float64 -> Float32 speedups:")
+    for N in Ns
+        bn32 = benchmark_name(N, :GPU, Float32)
+        bn64 = benchmark_name(N, :GPU, Float64)
+        t32  = TimerOutputs.time(timer[bn32])
+        t64  = TimerOutputs.time(timer[bn64])
+        @printf("%s: %.3f\n", benchmark_name(N), t64/t32)
+    end
+
+    println("\nCPU -> GPU speedsup:")
+    for N in Ns, ft in float_types
+        bn_cpu = benchmark_name(N, :CPU, ft)
+        bn_gpu = benchmark_name(N, :GPU, ft)
+        t_cpu  = TimerOutputs.time(timer[bn_cpu])
+        t_gpu  = TimerOutputs.time(timer[bn_gpu])
+        @printf("%s: %.3f\n", benchmark_name(N, ft), t_gpu/t_cpu)
+    end
+end