@glwagner 2019-03-08
That's cool!

Do you want to add these results (along with some information about the machines you used) to the documentation?

Do you think that we should allow people to plan multithreaded FFTs on the CPU? There's no reason not to use multithreading if its available.