@glwagner 2019-03-08
Also: where is the script? I can try to run the same on my GPU.